requirements.txt
# Instagram Story Downloader - VPS Version
# Python 3.9+ required

# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0

# HTTP Client
requests>=2.31.0
urllib3>=2.0.0

# WSGI Server (Production)
gunicorn>=21.0.0
gevent>=23.0.0

# Utilities
python-dotenv>=1.0.0
